+++ src/js/index.js
@@ -1,10 +1,11 @@
 import { stopFoit } from './modules/stopFoit';
 
 // All helper functions will be imported here, so that they can all be exported within one object.
-import { getBreakpoints, getCurrentScreenWidth } from './modules/breakpointHelper';
+import { getBreakpoints, getCurrentScreenWidth, isWithinBreakpoint } from './modules/breakpointHelper';
 
 export {
     stopFoit,
     getBreakpoints,
-    getCurrentScreenWidth
+    getCurrentScreenWidth,
+    isWithinBreakpoint
 };